Isolate Parser, combine it with Combinators in a top level reexpot module
