+++ json-parser.cabal
@@ -31,10 +31,11 @@ library
   import: shared-properties
   exposed-modules:
     Parser
-    Primitives
     Json
     Decoder
   other-modules:
+    Parser.Base
+    Parser.Combinators
   hs-source-dirs:
       lib
